+++ react/ProductKit.js
@@ -36,25 +36,30 @@ class ProductKit extends Component {
           type: 'boolean',
           title: 'editor.productKit.showListPrice',
           default: true,
+          isLayout: true,
         },
         showLabels: {
           type: 'boolean',
           title: 'editor.productKit.showLabels',
           default: false,
+          isLayout: true,
         },
         showInstallments: {
           type: 'boolean',
           title: 'editor.productKit.showInstallments',
           default: false,
+          isLayout: true,
         },
         showBadge: {
           type: 'boolean',
           title: 'editor.productKit.showBadge',
           default: false,
+          isLayout: true,
         },
         badgeText: showBadge ? {
           type: 'string',
           title: 'editor.productKit.badgeText',
+          isLayout: false,
         } : {},
       },
     }