+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `isLayout` property to getSchema of ProductKit component.
 
 ## [0.1.0] - 2018-7-6
-
 ### Added
 - PropTypes validation according to graphql data.
 - Initial implementation